web/requirements.txt
pandas
python-dotenv
flet
